+++ C/impl/sonLibList.c
@@ -97,7 +97,7 @@ void stList_set(stList *list, int32_t index, void *item) {
 static void *st_list_appendP(void *current, int32_t *currentSize, int32_t newSize, int32_t base) {
     assert(*currentSize <= newSize);
     void *new;
-    new = memcpy(st_malloc(base*newSize), current, base*(*currentSize));
+    new = memcpy(st_malloc(((int64_t)base)*newSize), current, ((int64_t)base)*(*currentSize));
     if(current != NULL) {
         free(current);
     }
